Disable CodeQL job and add conditional configurations
The CodeQL job in `codeql.yml` is now disabled with `if: false`.
The `runs-on` and `timeout-minutes` configurations are conditionally
set based on `matrix.language`. If the language is 'swift', it uses
'macos-latest' and a timeout of 120 minutes; otherwise, it uses
'ubuntu-latest' and a timeout of 360 minutes. The `permissions`
section remains unchanged.