Update SonarCloud project key and organization
The changes in the `sonarcloud.yml` file update the SonarCloud project key and organization. Specifically, the project key is changed from `lekomp_vigilant-umbrella` to `ffcoders_vigilant-umbrella`, and the organization is changed from `lekomp` to `ffcoders`. This update ensures that the SonarCloud analysis is associated with the correct project and organization.